+++ src/test/java/org/folio/rs/controller/TenantControllerTest.java
@@ -7,8 +7,8 @@ import io.zonky.test.db.AutoConfigureEmbeddedDatabase;
 import org.folio.rs.domain.AsyncFolioExecutionContext;
 import org.folio.spring.FolioExecutionContext;
 import org.folio.spring.FolioModuleMetadata;
+import org.folio.spring.controller.TenantController;
 import org.folio.spring.scope.FolioExecutionContextSetter;
-import org.folio.spring.scope.FolioExecutionScopeExecutionContextManager;
 import org.folio.tenant.domain.dto.TenantAttributes;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,12 +39,12 @@ public class TenantControllerTest {
 
   @Test
   void shouldDropDatabaseSchemaUponTenantDeletion() {
-    try (var context = new FolioExecutionContextSetter(AsyncFolioExecutionContext.builder()
+    try (var ignored = new FolioExecutionContextSetter(AsyncFolioExecutionContext.builder()
       .tenantId("test_tenant")
       .moduleMetadata(moduleMetadata)
       .okapiUrl(getOkapiUrl()).build())) {
       tenantController.postTenant(new TenantAttributes().moduleTo("mod_remote_storage"));
-      tenantController.deleteTenant("test_tenant");
+      tenantController.postTenant(new TenantAttributes().purge(true));
       verify(jdbcTemplate).execute(String.format(DROP_SCHEMA_QUERY, getSchemaName()));
     }
   }